feat(@rampant-core/plugin): Convert all log effects to proper strings.
When trying to just get the raw values of a source, I want to just be able to log out the result so I can start working on a new intent. It required a weird parsing step to stringify things and now we default to stringifying anything that is not already a string to make intent building easier.

dx(dependencies): Need proper is string check from lodash.

Closes #685