feat(@rampant/engine): Fill intents with reasonable defaults.
In order to speed up our ability to create intents for testing we want to minimize the amount of overhead that exists. One of these ways is to assume that an intent exists just needs to be run immediately when the intent is created/enabled/etc. with an intent filter.

test(@rampant/ai): Remove extraneous source and log message parsing.

Closes #564